+++ cleo/__init__.py
@@ -2,6 +2,7 @@
 
 from .application import Application
 from .commands import Command
+from .helpers import argument, option
 from .testers import ApplicationTester, CommandTester